requirements.txt
streamlit>=1.28
psycopg2-binary>=2.9
pandas>=2.0
openpyxl>=3.1
fpdf2>=2.7
